feat: diagnostic styles and configuration of components

+++ client/src/pages/Logout.jsx
@@ -1,33 +1,42 @@
-import React, { Component } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { useNavigate } from "react-router-dom";
-import axios from 'axios'
+import axios from 'axios';
 
-// APis
+// APIs
 import { LogoutUser } from '../api/logout.api';
 
-export const Logout = async () => {
+const Logout = () => {
     const navigate = useNavigate();
-    const refresh = { "refresh_token": localStorage.getItem('refresh_token') };
-    try {
-        // Llamada a la API
-        const response = await LogoutUser(refresh);
-        console.log("Respuesta del servidor Logout:", response?.status);
-        localStorage.clear();
-        axios.defaults.headers.common['Authorization'] = null;
-        navigate("/login");
+    const hasLoggedOut = useRef(false); // Indicador para evitar múltiples llamadas
 
-    } catch (error) {
-        console.error("Error al cerrar sesión:", error);
-        window.location.href = '/'
-        alert("Ocurrió un error al cerrar sesión. Intente nuevamente..");
-    }
+    useEffect(() => {
+        const performLogout = async () => {
+            if (hasLoggedOut.current) return; // Si ya se ejecutó, no hacer nada
+            hasLoggedOut.current = true; // Marcar como ejecutado
+
+            const refresh = { "refresh_token": localStorage.getItem('refresh_token') };
+            try {
+                // Llamada a la API
+                const response = await LogoutUser(refresh);
+                console.log("Respuesta del servidor Logout:", response?.status);
+                localStorage.clear();
+                axios.defaults.headers.common['Authorization'] = null;
+                navigate("/login"); // Redirige al login
+            } catch (error) {
+                console.error("Error al cerrar sesión:", error);
+                alert("Ocurrió un error al cerrar sesión. Intente nuevamente.");
+                navigate("/"); // Redirige al inicio en caso de error
+            }
+        };
+
+        performLogout();
+    }, [navigate]); // Se ejecuta solo una vez al montar el componente
 
     return (
-        <>
-            <div></div>
-        </>
-    )
-
-}
+        <div>
+            <p>Cerrando sesión...</p>
+        </div>
+    );
+};
 
 export default Logout;
